Add app.rb
